change: refactor kernel computation method
